pyproject.toml
[project]
name = "coordination-agent"
version = "0.1.0"
description = "coordination agent demo using Agent Development Kit"
authors = [{ name = "Nathan Fu", email = "me@nathanfu.dev" }]
requires-python = ">=3.11,<3.14"
readme = "README.md"
license = "Apache License 2.0"
dependencies = [
    "google-cloud-aiplatform[agent_engines, adk]",
    "ortools>=9.12.4544,<10",
    "litellm>=1.69.0,<2",
    "google-adk>=1.5.0,<2",
    "pydantic>=2.11.5",
]

[dependency-groups]
dev = [
    "pydantic-settings>=2.8.1,<3",
    "flake8-pyproject>=1.2.3,<2",
    "pylint>=3.3.6,<4",
    "pyink>=24.10.1,<25",
    "google-adk[eval]>=1.5.0",
    "pytest-asyncio>=1.0.0",
    "pytest>=8.4.0",
]

[tool.uv]

[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
testpaths = ["tests/", "eval/"]
markers = [
    "unit"
]
log_level="ERROR"
log_cli=false
log_auto_indent=true
log_cli_date_format="%Y-%m-%d %H:%M:%S"
log_cli_format ="[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = [
    "ignore::UserWarning",
]

[tool.pyink]
line-length=80
pyink-indentation=4
pyink-use-majority-quotes = true
